Ant build.xml created and project now compatible with Tomcat 6 and Java 1.6

+++ workspace/richturner/DataLogger/src/org/openremote/datalogger/connector/DataConnector.java
@@ -20,7 +20,6 @@
 package org.openremote.datalogger.connector;
 
 import java.util.Date;
-import java.util.HashMap;
 import java.util.Set;
 import org.openremote.datalogger.exception.DataConnectorException;
 import org.openremote.datalogger.exception.DataSecurityException;
@@ -42,7 +41,7 @@ public interface DataConnector {
 	// Set multiple values for a particular sensor ID and API Key
 	void addSensorValues(String apiKey, String sensorName, Set<SensorValue> values, String currentValue) throws DataSecurityException, DataConnectorException;
 	
-	float getAverageSensorValue(String apiKey, String sensorName, Date fromTime, Date toTime) throws DataSecurityException, DataConnectorException;
+	Double getAverageSensorValue(String apiKey, String sensorName, Date fromTime, Date toTime) throws DataSecurityException, DataConnectorException;
 	
 	// Cleanup the data connector
 	void destroy();
